Add decimal functionality

+++ Calculator.rthring/Calculator.rthring/Program.cs
@@ -1,13 +1,13 @@
-﻿int num1 = 0; int num2 = 0;
+﻿double num1 = 0; double num2 = 0;
 
 Console.WriteLine("Console Calculator in C#\r");
 Console.WriteLine("------------------------\n");
 
 Console.WriteLine("Type a number, and then press Enter");
-num1 = Convert.ToInt32(Console.ReadLine());
+num1 = Convert.ToDouble(Console.ReadLine());
 
 Console.WriteLine("Type another number, and then press Enter");
-num2 = Convert.ToInt32(Console.ReadLine());
+num2 = Convert.ToDouble(Console.ReadLine());
 
 Console.WriteLine("Choose an option from the following list:");
 Console.WriteLine("\ta - Add");
